Remove partial constructors
Both socket addresses need to be known before starting emulation.

If the receiver address would be unknown, the emulator does not know
where to send the packet to. However, if sender sits behind NAT router,
the message will never leave the JVM when routed through this emulator,
which also implies that no entry will be created in the router's
translation table. Packets coming from the receiver will thus get
dropped by the router, resulting in a "connection deadlock".

If the sender address is unknown, the emulator can easily set it by
inspecting the incoming packet. However, to avoid complicating things,
the constructor now requires both addresses, under the assumption that
an emulator sitting on the sender side has access to the sender address.